Reworked print_devices_and_exit() to not mangle the screen.
If you type "minipro -l" at the command line, what's supposed to happen
is that minipro pipes a list of devices to the less(1) command.  This
works, but commands used in less(1) are used (particularly something
like "10p", which scrolls down to the ten-percent mark) result in both
programs aborting and the screen is left in a mangled state.  This
modification makes minipro write its list of devices to a temporary
file, upon which less(1) is invoked.  Now there is no more mangling or
complaints of "waiting for output".